+++ sample-code/spring-app/src/main/java/com/sap/ai/sdk/app/controllers/OrchestrationController.java
@@ -148,13 +148,14 @@ class OrchestrationController {
     return response.getContent();
   }
 
-  @GetMapping("/outputFiltering/{policy}")
+  @GetMapping("/outputFiltering/{policy}/{isProtected}")
   @Nonnull
   Object outputFiltering(
       @Nullable @RequestParam(value = "format", required = false) final String format,
-      @Nonnull @PathVariable("policy") final AzureFilterThreshold policy) {
+      @Nonnull @PathVariable("policy") final AzureFilterThreshold policy,
+      @Nonnull @PathVariable("isProtected") final Boolean isProtected) {
 
-    final var response = service.outputFiltering(policy);
+    final var response = service.outputFiltering(policy, isProtected);
 
     final String content;
     try {
@@ -392,6 +393,16 @@ class OrchestrationController {
   @Nonnull
   Object configFromRegistry(@RequestParam(value = "format", required = false) final String format) {
     final var response = service.executeConfigFromReference();
+    if ("json".equals(format)) {
+      return response;
+    }
+    return response.getContent();
+  }
+
+  @GetMapping("/completionWithFallback")
+  Object completionWithFallback(
+      @Nullable @RequestParam(value = "format", required = false) final String format) {
+    final var response = service.completionWithFallback("HelloWorld!");
     if ("json".equals(format)) {
       return response;
     }