+++ sample-code/spring-app/src/test/java/com/sap/ai/sdk/app/controllers/OrchestrationTest.java
@@ -273,7 +273,7 @@ class OrchestrationTest {
   @Test
   void testOutputFilteringStrict() {
     var policy = AzureFilterThreshold.ALLOW_SAFE;
-    var response = service.outputFiltering(policy);
+    var response = service.outputFiltering(policy, true);
 
     assertThatThrownBy(response::getContent)
         .hasMessageContaining("Content filter filtered the output.")
@@ -286,6 +286,7 @@ class OrchestrationTest {
               assertThat(actualAzureContentSafety.getSelfHarm()).isEqualTo(NUMBER_0);
               assertThat(actualAzureContentSafety.getSexual()).isEqualTo(NUMBER_0);
               assertThat(actualAzureContentSafety.getHate()).isEqualTo(NUMBER_0);
+              assertThat(actualAzureContentSafety.isProtectedMaterialCode()).isFalse();
             });
   }
 
@@ -293,7 +294,7 @@ class OrchestrationTest {
   void testOutputFilteringLenient() {
     var policy = AzureFilterThreshold.ALLOW_ALL;
 
-    var response = service.outputFiltering(policy);
+    var response = service.outputFiltering(policy, false);
 
     assertThat(response.getChoice().getFinishReason()).isEqualTo("stop");
     assertThat(response.getContent()).isNotEmpty();
@@ -554,5 +555,41 @@ class OrchestrationTest {
     val result = service.executeConfigFromReference();
     val choices = (result.getOriginalResponse().getFinalResult()).getChoices();
     assertThat(choices.get(0).getMessage().getContent()).isNotEmpty();
+  }
+
+  @Test
+  void testCompletionWithFallback() {
+    val result = service.completionWithFallback("HelloWorld!");
+
+    assertThat(result).isNotNull();
+    assertThat(result.getContent()).isNotEmpty();
+    assertThat(result.getOriginalResponse().getIntermediateFailures().size()).isEqualTo(2);
+    assertThat(result.getOriginalResponse().getIntermediateFailures().get(0).getMessage())
+        .contains("Model broken_name not supported.");
+    assertThat(result.getOriginalResponse().getIntermediateFailures().get(0).getCode())
+        .isEqualTo(400);
+    assertThat(result.getOriginalResponse().getFinalResult().getChoices().get(0).getFinishReason())
+        .contains("stop");
+  }
+
+  @Test
+  void testCompletionWithFallbackStreaming() {
+    final var stream = service.streamCompletionWithFallback("HelloWorld!");
+    val filledDeltaCount = new AtomicInteger(0);
+    stream.forEach(
+        delta -> {
+          log.info("delta: {}", delta);
+          if (!delta.isEmpty()) {
+            filledDeltaCount.incrementAndGet();
+          }
+        });
+    assertThat(filledDeltaCount.get()).isGreaterThan(0);
+  }
+
+  @Test
+  void testCompletionWithFallbackAllFail() {
+    assertThatThrownBy(() -> service.completionWithFallbackAllFail("HelloWorld!"))
+        .isInstanceOf(OrchestrationClientException.class)
+        .hasMessageContaining("Model broken_name_2 not supported.");
   }
 }